agents/content-agent/requirements.txt
crewai
crewai_tools
google-cloud-aiplatform
google-cloud-firestore
google-cloud-pubsub
pydantic
python-dotenv
requests
python-wordpress-xmlrpc